+++ adapters/samsung/__init__.py
@@ -1,6 +1,7 @@
 from adapters.contact_adapter import ContactAdapter
 from adapters.on_off_switch_adapter import OnOffSwitchAdapter
 from adapters.generic.motion_temp_sensor import MotionAndTemperatureSensorAdapter
+from adapters.generic.on_off_kwh import OnOffKwhAdapter
 from adapters.samsung.sensor_arrival import SensorArrival
 from adapters.samsung.sensor_door import SmartThingsDoorSensor
 
@@ -17,6 +18,8 @@ samsung_adapters = {
     'IM6001-MPP01': ContactAdapter,     # SmartThings Multipurpose sensor (2018 model)
     '3300-S': SmartThingsDoorSensor,    # SmartThings Door sensor
     '3321-S': SmartThingsDoorSensor,    # SmartThings Multi Sensor (2015 model)
-    'STS-OUT-US-2': OnOffSwitchAdapter, # SmartThings Outlet
+    'STS-OUT-US-2': OnOffKwhAdapter,    # SmartThings Zigbee smart plug with power meter
     'F-APP-UK-V2': OnOffSwitchAdapter,  # SmartThings Outlet
+    'GP-WOU019BBDWG': OnOffKwhAdapter,  # SmartThings Outlet with power meter
+    'F-APP-UK-V2': OnOffKwhAdapter,     # SmartThings Zigbee Outlet UK with power meter
 }